LayoutTests/fast/harness/perftests/runs-per-second-log.html

<!DOCTYPE html>
<html>
<body>
<p id="description">This test verifies PerfTestRunner.runPerSecond() outputs log as expected.</p>
<div id="console"></div>
<script src="../../../../PerformanceTests/resources/runner.js"></script>
<script type="text/javascript">

var logLines = [];
var initial = true;
var runs = 0;

var iteration = 0;
var currentTime = 0;

PerfTestRunner.now = function () {
    iteration++;
    if (iteration % 2 == 0)
        currentTime += 1000 / runs;
    return currentTime;
}

PerfTestRunner.storeHeapResults = function () { }

var printStatistics = PerfTestRunner.printStatistics;
PerfTestRunner.printStatistics = function (statistics, title) {
    if (statistics.unit == 'bytes')
        return;
    statistics.stdev = statistics.stdev.toPrecision(3);
    return printStatistics.call(PerfTestRunner, statistics, title);
}

PerfTestRunner.measureRunsPerSecond({
    setup: function () {
        runs++;
        // Avoid 3, 6, 7, 9.
        if (runs == 6)
            runs += 2;
        else if (!(runs % 3))
            runs++;
    },
    run: function () { },
    runCount: 5,
});

</script>
</body>
</html>
